==> r_scripts/intercoder_agreement/03_intercoder_agreement.Rmd <==
---
title: "Inter-rater agreement"
output: html_document
date: "2025-03-10"
---

```{r setup, include=FALSE}

# Load packages
library(here)

# Load function to estimate inter-coder agreement between 3 and 2 coders
source(here("r_scripts", "intercoder_agreement", "02_intercoder_agreement_function.R"))

```


# Inter-rater agreement between 3 coders: Fleiss' kappa
```{r echo=FALSE}

# Calculate Fleiss' kappa by coded variable
agreement_3_coders <- calculate_kappa3_from_excel(here("data/disagreements3.xlsx"))
print(agreement_3_coders)

```


# Average Fleiss' kappa
```{r echo=FALSE}

# Calculate average Fleiss' kappa for "support_hypothesis", "type_hypothesis" and
# "statistical_result", "type_effect"
mean(agreement_3_coders[c(16, 17, 19, 20), 2]) 

```

# Inter-rater agreement between 2 coders: Cohen's Kappa
```{r echo=FALSE}

# Calculate Cohen's kappa by coded variable
agreement_2_coders <- calculate_kappa2_from_excel(here("data/disagreements2.xlsx"))
print(agreement_2_coders)

```


# Average Cohen's Kappa
```{r echo=false}

# Calculate average Cohen's Kappa for "support_hypothesis", "type_hypothesis",
# "statistical_result" and "type_effect"
mean(agreement_2_coders[c(16, 17, 19, 20), 2])

```
